\encoding{utf8} 
\name{mSimCC-internal}
\alias{mSimCC-internal}
\alias{byAgeGroup2}
\alias{byAgeGroup3}
\alias{Cmicrosim}

\title{Internal mSimCC functions}
\description{Internal mSimCC functions}
\usage{
byAgeGroup2(data)
byAgeGroup3(data)
}

\details{These functions are not to be called by the user}
\seealso{
\code{\link{mSimCC-package}}, \code{\link{bCohort}}, \code{\link{microsim}}, \code{\link{costs}}, \code{\link{le}},
\code{\link{plotCIN1Incidence}}, \code{\link{plotCIN2Incidence}}, \code{\link{plotCIN3Incidence}},
\code{\link{plotIncidence}}, \code{\link{plotMortality}}, \code{\link{plotPrevalence}},
\code{\link{qalys}}, \code{\link{yls}}
}
\keyword{ internal }
