\encoding{utf8}
\name{mSimCC-package}
\alias{mSimCC-package}
\alias{mSimCC}
\docType{package}
\title{
Microsimulation model for cervical cancer prevention
}
\description{
Microsimulation model to reproduce natural history of cervical cancer and cost-effectiveness evaluation of prevention strategies.
}
\details{
\tabular{ll}{
Package: \tab mSimCC\cr
Type: \tab Package\cr
Version: \tab 0.0.3\cr
Date: \tab 2023-08-21\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}
}
\author{
David Moriña (Universitat de Barcelona), Pedro Puig (Universitat Autònoma de Barcelona) and Mireia Diaz (Institut Català d'Oncologia)

Mantainer: David Moriña Soler <dmorina@ub.edu>
}

\seealso{
\code{\link{mSimCC-package}}, \code{\link{bCohort}}, \code{\link{microsim}}, \code{\link{costs}}, \code{\link{le}},
\code{\link{plotCIN1Incidence}}, \code{\link{plotCIN2Incidence}}, \code{\link{plotCIN3Incidence}},
\code{\link{plotIncidence}}, \code{\link{plotMortality}}, \code{\link{plotPrevalence}},
\code{\link{qalys}}, \code{\link{yls}}
}

\references{
Georgalis L, de Sanjosé S, Esnaola M, Bosch F X, Diaz M. Present and future of cervical cancer prevention in Spain: a cost-effectiveness analysis. European Journal of Cancer Prevention 2016;\bold{25}(5):430-439.

Moriña D, de Sanjosé S, Diaz M. Impact of model calibration on cost-effectiveness analysis of cervical cancer prevention 2017;\bold{7}.
}
\keyword{ mSimCC }
