\encoding{utf8} 
\name{probs}
\docType{data}
\alias{probs}
\title{Transition probabilities matrix for Spanish population}
\description{
  This data corresponds to a transition probabilities matrix calibrated for the Spanish population.  
}
\usage{probs}
\format{A data frame with 180 rows and 13 columns.}

\keyword{ datasets }